refactor: isolar contextos de carrinho e pedido e limpar pacotes antigos
