+++ src/main/java/com/natodev/ecommerce/domain/cart/domain/entity/ItemCarrinho.java
@@ -1,4 +1,4 @@
-package com.natodev.ecommerce.infrastructure.entity;
+package com.natodev.ecommerce.domain.cart.domain.entity;
 
 import com.natodev.ecommerce.domain.product.domain.entity.Produto;
 import jakarta.persistence.*;
@@ -13,29 +13,25 @@ import java.util.UUID;
 @AllArgsConstructor
 @NoArgsConstructor
 @Entity
-@Table(name = "item_pedido")
-public class ItemPedido {
+@Table(name = "item_carrinho")
+public class ItemCarrinho {
 
     @Id
+    @Column(name = "item_carrinho_id")
     @GeneratedValue(strategy = GenerationType.UUID)
-    @Column(name = "item_pedido_id")
-    private UUID itemPedidoId;
+    private UUID itemCarrinhoId;
 
     @ManyToOne
-    @JoinColumn (name = "pedido_id")
-    private Pedido pedido;
+    @JoinColumn(name = "carrinho_id")
+    private Carrinho carrinho;
 
     @ManyToOne
-    @JoinColumn (name = "produto_id")
+    @JoinColumn(name = "produto_id")
     private Produto produto;
 
-    @Column(name = "quantidade")
+    @Column (name = "quantidade")
     private Integer quantidade;
 
-    @Column(name = "preco_unitario")
+    @Column (name = "preco_unitario")
     private BigDecimal precoUnitario;
-
-    @Column(name = "sub_total")
-    private BigDecimal subTotal;
-
 }
