+++ src/main/java/com/natodev/ecommerce/domain/cart/domain/repository/ItemCarrinhoRepository.java
@@ -1,7 +1,7 @@
-package com.natodev.ecommerce.infrastructure.repository;
+package com.natodev.ecommerce.domain.cart.domain.repository;
 
-import com.natodev.ecommerce.infrastructure.entity.Carrinho;
-import com.natodev.ecommerce.infrastructure.entity.ItemCarrinho;
+import com.natodev.ecommerce.domain.cart.domain.entity.Carrinho;
+import com.natodev.ecommerce.domain.cart.domain.entity.ItemCarrinho;
 import com.natodev.ecommerce.domain.product.domain.entity.Produto;
 import org.springframework.data.jpa.repository.JpaRepository;
