+++ src/main/java/com/natodev/ecommerce/domain/cart/presentation/dto/response/CarrinhoResponseDTO.java
@@ -1,4 +1,4 @@
-package com.natodev.ecommerce.controller.dto.response;
+package com.natodev.ecommerce.domain.cart.presentation.dto.response;
 
 import java.time.LocalDate;
 import java.util.UUID;
