+++ src/main/java/com/natodev/ecommerce/domain/order/presentation/dto/request/ItemPedidoRequestDTO.java
@@ -1,17 +1,16 @@
-package com.natodev.ecommerce.controller.dto.request;
+package com.natodev.ecommerce.domain.order.presentation.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 
 import java.util.UUID;
 
-public record ItemCarrinhoRequestDTO(
+public record ItemPedidoRequestDTO(
 
-        @NotNull(message = "Item precisa de um produto")
+        @NotNull(message = "Item pedido precisa de um produto")
         UUID produtoId,
 
         @NotNull(message = "Quantidade é obrigatória")
         @Positive(message = "Quantidade deve ser maior que zero")
         Integer quantidade
-
 ) {}
