+++ src/main/java/com/natodev/ecommerce/domain/order/presentation/dto/response/ItemPedidoResponseDTO.java
@@ -1,4 +1,4 @@
-package com.natodev.ecommerce.controller.dto.response;
+package com.natodev.ecommerce.domain.order.presentation.dto.response;
 
 import java.math.BigDecimal;
 import java.util.UUID;
