+++ src/main/java/com/natodev/ecommerce/domain/user/domain/entity/Usuario.java
@@ -1,7 +1,7 @@
 package com.natodev.ecommerce.domain.user.domain.entity;
 
-import com.natodev.ecommerce.infrastructure.entity.Carrinho;
-import com.natodev.ecommerce.infrastructure.entity.Pedido;
+import com.natodev.ecommerce.domain.cart.domain.entity.Carrinho;
+import com.natodev.ecommerce.domain.order.domain.entity.Pedido;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
